Animal party and wild animals
Wild animals added to the world so that the enemy doesn't have to hard coded. Also scripts to store and initialise the animals in the players party and to check and send out any healthy animals if one faints. Also added tiger and tiger cub sprites to make another animal.